closes #20 : Added ValidateUserCredentials Method in AuthService (#62)

+++ src/main/java/com/bank/mapper/UserMapper.java
@@ -38,7 +38,7 @@ public class UserMapper {
         }
 
         UserDTO userDTO=new UserDTO();
-        userDTO.setId((long) user.getId());
+        userDTO.setId( user.getId());
         userDTO.setUsername(user.getUsername());
         userDTO.setFullName(user.getFullName());
         userDTO.setEmail(user.getEmail());
